add validation and pagination

+++ ELodowka.Api/Controllers/IngredientController.cs
@@ -1,7 +1,9 @@
 ﻿using ELodowka.Api.Common.DTOs.Ingredients;
+using ELodowka.Api.Common.Exceptions;
 using ELodowka.Api.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
 
 namespace ELodowka.Api.Controllers;
 
@@ -17,15 +19,39 @@ public class IngredientController : ControllerBase
         _ingredientService = ingredientService;
     }
 
-    [HttpGet]
-    public async Task<List<IngredientDto>> GetMany()
+    [HttpGet("{page:int}/{itemsPerPage:int}")]
+    public async Task<ActionResult<List<IngredientDto>>> GetMany([FromRoute] int page, [FromRoute] int itemsPerPage)
     {
-        return await _ingredientService.GetMany();
+        var ingredients = await _ingredientService.GetMany();
+        var totalCount = ingredients.Count;
+        var totalPages = (int)Math.Ceiling((double)totalCount / itemsPerPage);
+
+        if (page < 1 || page > totalPages)
+        {
+            return BadRequest("Invalid page number");
+        }
+
+        var skip = (page - 1) * itemsPerPage;
+        var take = itemsPerPage;
+
+        var result = ingredients.Skip(skip).Take(take).ToList();
+
+        return Ok(new 
+        {
+            totalPages,
+            totalCount,
+            result
+        });
+        
     }
 
     [HttpPost]
     public async Task<ActionResult> Add([FromBody] IngredientDto model)
     {
+        if (model == null) throw new ModelErrorException();
+        if (model.Name.Length < 1) throw new ModelErrorException();
+        if (model.Type.Length < 1) throw new ModelErrorException();
+
         await _ingredientService.Add(model);
         return Ok();
     }
@@ -33,6 +59,10 @@ public class IngredientController : ControllerBase
     [HttpPut("{id:long}")]
     public async Task<ActionResult> Update([FromRoute] long id, [FromBody] IngredientDto model)
     {
+        if (id < 0) throw new ModelErrorException();
+        if (model == null) throw new ModelErrorException();
+        if (model.Name.Length < 1) throw new ModelErrorException();
+        if (model.Type.Length < 1) throw new ModelErrorException();
         await _ingredientService.Update(id, model);
         return Ok();
     }
@@ -40,12 +70,14 @@ public class IngredientController : ControllerBase
     [HttpGet("{id:long}")]
     public async Task<ActionResult<IngredientDto>> Get([FromRoute] long id)
     {
+        if (id < 0) throw new ModelErrorException();
         return await _ingredientService.Get(id);
     }
 
     [HttpDelete("{id:long}")]
     public async Task<ActionResult> Delete([FromRoute] long id)
     {
+        if (id < 0) throw new ModelErrorException();
         await _ingredientService.Delete(id);
         return Ok();
     }
